Fix serialization tests
